Moved functions primarily used for operand reading into their own file.
