+++ src/parser.rs
@@ -3,3 +3,4 @@
 pub mod parser_main;
 pub mod parser_preprocessing;
 pub mod parser_structs_and_enums;
+pub mod operand_reading;
